requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
certifi==2025.10.5
charset-normalizer==3.4.4
distro==1.9.0
Django==5.2.6
exceptiongroup==1.3.0
faiss-cpu==1.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.0
joblib==1.5.2
numpy==2.2.6
openai==2.3.0
packaging==25.0
pdf2image==1.17.0
pillow==12.0.0
psycopg==3.2.10
psycopg-binary==3.2.10
pydantic==2.12.2
pydantic_core==2.41.4
PyMuPDF==1.26.5
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.1.1
regex==2025.10.22
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
